refactor: decompose analysis_widgets and folder_tool_file_ops god modules
Break down two large monolith modules per Pragmatic Programmer assessment:

analysis_widgets.py (1,558 lines -> 5 focused modules):
- statistical_widgets.py (529): VariableSelector, PCA, ANOVA, Regression
- visualization_widgets.py (505): SurfacePlot, NeuralNetwork, ScriptGenerator
- plot_dialogs.py (336): Contour, Heatmap, FilterComparison dialogs
- chart_style_panel.py (246): Per-series chart style controls
- analysis_widgets.py (112): Facade + AnalysisPanel compositor

folder_tool_file_ops.py (1,179 lines -> 5 focused modules):
- file_validation.py (207): Filter/size validation, path organization
- backup_copy.py (402): Backup creation, safe copy with retry
- archive_ops.py (213): Archive extraction with validation
- folder_ops.py (409): Combine, deduplicate, flatten, prune
- folder_tool_file_ops.py (30): Composite mixin facade

Key design decisions:
- Backward compatibility preserved via facade pattern
- All existing imports continue to work unchanged
- Mixin composition for folder_tool maintains MRO integrity
